+++ kamehouse-groot/public/kame-house-groot/api/v1/auth/login.php
@@ -40,19 +40,6 @@
   }
 
   /**
-   * Start a new session.
-   */
-  function initiateSession($username) {
-    ini_set('session.gc_maxlifetime', 0);
-    session_set_cookie_params(0);
-    session_start();
-    session_regenerate_id();
-    $_SESSION['logged-in'] = true;
-    $_SESSION['username'] = $username;
-    unlockSession();
-  }
-
-  /**
    * Redirect after successful login.
    */
   function redirectLoginSuccess() {